disable token caching by default

+++ vault.sh
@@ -93,8 +93,8 @@ get_new_password () {
 get_token () {
     TOKEN_DIR=".vault"
     TOKEN_FILE="${TOKEN_DIR}/token"
-    # [ "$NO_CACHE_TOKEN" ] && rm -rf ${TOKEN_DIR} &> /dev/null
-    rm -rf ${TOKEN_DIR} &> /dev/null
+    [ "$NO_CACHE_TOKEN" ] && rm -rf ${TOKEN_DIR} &> /dev/null
+    #rm -rf ${TOKEN_DIR} &> /dev/null
     mkdir -p ${TOKEN_DIR} &> /dev/null
 
     get_vault_url
@@ -204,6 +204,7 @@ while [[ "$#" -gt 0 ]]; do
         list_users            ) ACTION="list_users"               ;;
         add_service_policy    ) ACTION="add_service_policy"       ;;
         add_service_user      ) ACTION="add_service_user"         ;;
+        --no-cache-token      ) NO_CACHE_TOKEN="true"             ;;
         -w|--website|--url    ) VAULT_URL="$2"           ; shift  ;;
         -a|--auth-method      ) VAULT_AUTH_METHOD="$2"   ; shift  ;;
         -u|--username         ) VAULT_USERNAME="$2"      ; shift  ;;
